+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/OpenTelemetryProtocolExporterEventSource.cs
@@ -16,7 +16,6 @@
 
 using System;
 using System.Diagnostics.Tracing;
-using System.Security;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation
@@ -25,24 +24,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation
     internal class OpenTelemetryProtocolExporterEventSource : EventSource
     {
         public static readonly OpenTelemetryProtocolExporterEventSource Log = new OpenTelemetryProtocolExporterEventSource();
-
-        [NonEvent]
-        public void MissingPermissionsToReadEnvironmentVariable(SecurityException ex)
-        {
-            if (this.IsEnabled(EventLevel.Warning, EventKeywords.All))
-            {
-                this.MissingPermissionsToReadEnvironmentVariable(ex.ToInvariantString());
-            }
-        }
-
-        [NonEvent]
-        public void FailedToConvertToProtoDefinitionError(Exception ex)
-        {
-            if (Log.IsEnabled(EventLevel.Error, EventKeywords.All))
-            {
-                this.FailedToConvertToProtoDefinitionError(ex.ToInvariantString());
-            }
-        }
 
         [NonEvent]
         public void FailedToReachCollector(Exception ex)
@@ -60,12 +41,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation
             {
                 this.ExportMethodException(ex.ToInvariantString());
             }
-        }
-
-        [Event(1, Message = "Exporter failed to convert SpanData content into gRPC proto definition. Data will not be sent. Exception: {0}", Level = EventLevel.Error)]
-        public void FailedToConvertToProtoDefinitionError(string ex)
-        {
-            this.WriteEvent(1, ex);
         }
 
         [Event(2, Message = "Exporter failed send data to collector. Data will not be sent. Exception: {0}", Level = EventLevel.Error)]
@@ -90,18 +65,6 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation
         public void CouldNotTranslateMetric(string className, string methodName)
         {
             this.WriteEvent(5, className, methodName);
-        }
-
-        [Event(6, Message = "Failed to parse environment variable: '{0}', value: '{1}'.", Level = EventLevel.Warning)]
-        public void FailedToParseEnvironmentVariable(string name, string value)
-        {
-            this.WriteEvent(6, name, value);
-        }
-
-        [Event(7, Message = "Missing permissions to read environment variable: '{0}'", Level = EventLevel.Warning)]
-        public void MissingPermissionsToReadEnvironmentVariable(string exception)
-        {
-            this.WriteEvent(7, exception);
         }
 
         [Event(8, Message = "Unsupported value for protocol '{0}' is configured, default protocol 'grpc' will be used.", Level = EventLevel.Warning)]