+++ src/OpenTelemetry/Trace/BatchExportActivityProcessorOptions.cs
@@ -16,7 +16,6 @@
 
 using System;
 using System.Diagnostics;
-using System.Security;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Trace
@@ -27,7 +26,7 @@ namespace OpenTelemetry.Trace
     /// environment variables are parsed during object construction.
     /// </summary>
     /// <remarks>
-    /// The constructor throws <see cref="ArgumentException"/> if it fails to parse
+    /// The constructor throws <see cref="FormatException"/> if it fails to parse
     /// any of the supported environment variables.
     /// </remarks>
     public class BatchExportActivityProcessorOptions : BatchExportProcessorOptions<Activity>
@@ -44,55 +43,25 @@ namespace OpenTelemetry.Trace
         {
             int value;
 
-            if (LoadEnvVarInt(ExporterTimeoutEnvVarKey, out value))
+            if (EnvironmentVariableHelper.LoadNumeric(ExporterTimeoutEnvVarKey, out value))
             {
                 this.ExporterTimeoutMilliseconds = value;
             }
 
-            if (LoadEnvVarInt(MaxExportBatchSizeEnvVarKey, out value))
+            if (EnvironmentVariableHelper.LoadNumeric(MaxExportBatchSizeEnvVarKey, out value))
             {
                 this.MaxExportBatchSize = value;
             }
 
-            if (LoadEnvVarInt(MaxQueueSizeEnvVarKey, out value))
+            if (EnvironmentVariableHelper.LoadNumeric(MaxQueueSizeEnvVarKey, out value))
             {
                 this.MaxQueueSize = value;
             }
 
-            if (LoadEnvVarInt(ScheduledDelayEnvVarKey, out value))
+            if (EnvironmentVariableHelper.LoadNumeric(ScheduledDelayEnvVarKey, out value))
             {
                 this.ScheduledDelayMilliseconds = value;
             }
-        }
-
-        private static bool LoadEnvVarInt(string envVarKey, out int result)
-        {
-            result = 0;
-
-            string value;
-            try
-            {
-                value = Environment.GetEnvironmentVariable(envVarKey);
-            }
-            catch (SecurityException ex)
-            {
-                // The caller does not have the required permission to
-                // retrieve the value of an environment variable from the current process.
-                OpenTelemetrySdkEventSource.Log.MissingPermissionsToReadEnvironmentVariable(ex);
-                return false;
-            }
-
-            if (string.IsNullOrEmpty(value))
-            {
-                return false;
-            }
-
-            if (!int.TryParse(value, out result))
-            {
-                throw new ArgumentException($"{envVarKey} environment variable has an invalid value: '${value}'");
-            }
-
-            return true;
         }
     }
 }