+++ test/Benchmarks/Exporter/OtlpHttpExporterBenchmarks.cs
@@ -14,16 +14,18 @@
 // limitations under the License.
 // </copyright>
 
+extern alias OpenTelemetryProtocol;
+
 using System;
 using System.Diagnostics;
 using System.IO;
 using BenchmarkDotNet.Attributes;
 using Benchmarks.Helper;
 using OpenTelemetry;
-using OpenTelemetry.Exporter;
-using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Tests;
+using OpenTelemetryProtocol::OpenTelemetry.Exporter;
+using OpenTelemetryProtocol::OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
 
 namespace Benchmarks.Exporter
 {